requirements.txt
fastapi
uvicorn
pydantic
yt-dlp
python-dotenv